+++ collective/venue/interfaces.py
@@ -12,19 +12,17 @@ from plone.supermodel import model
 from zope import schema
 from zope.interface import Interface
 from zope.interface import provider
+from plone.app.vocabularies.catalog import CatalogSource
 
 
 @provider(IFormFieldProvider)
 class IVenue(model.Schema):
     """Marker schema interface for Venue types."""
+
     notes = RichText(
-        title=_(
-            u'label_notes',
-            default=u'Notes'
-        ),
+        title=_(u'label_notes', default=u'Notes'),
         description=_(
-            u'help_notes',
-            default=u'Additional notes for the address.'
+            u'help_notes', default=u'Additional notes for the address.'
         ),
         required=False,
     )
@@ -36,15 +34,12 @@ class IVenueSettings(Interface):
     """
 
     search_base = schema.Choice(
-        title=_(
-            u'label_search_base',
-            default=u'Location Search Base'
-        ),
+        title=_(u'label_search_base', default=u'Location Search Base'),
         description=_(
             u'help_search_base',
             u"Path, from which venue types should be searched. Useful for "
             u"lineage multisites to seperate main from childsite venue "
-            u"folders. Keep empty to search anywhere."
+            u"folders. Keep empty to search anywhere.",
         ),
         required=False,
         default='',
@@ -56,20 +51,18 @@ class IVenueSettings(Interface):
         pattern_options={
             'selectableTypes': ['Folder'],  # better: is_folderish
             'basePath': get_site,
-        }
+        },
     )
 
     default_venue = schema.Choice(
-        title=_(
-            u'label_default_venue',
-            default=u'Default Location'
-        ),
+        title=_(u'label_default_venue', default=u'Default Location'),
         description=_(
-            u'help_default_venue',
-            u"Default location to be used in events."),
+            u'help_default_venue', u"Default location to be used in events."
+        ),
         required=False,
         default='',
         vocabulary='plone.app.vocabularies.Catalog',
+        # source=CatalogSource(portal_type=""),
     )
     form.widget(
         'default_venue',
@@ -77,17 +70,15 @@ class IVenueSettings(Interface):
         pattern_options={
             'selectableTypes': ['Venue'],
             'basePath': get_base_path,
-        }
+        },
     )
 
     default_organizer = schema.Choice(
-        title=_(
-            u'label_default_organizer',
-            default=u'Default Organizer'
-        ),
+        title=_(u'label_default_organizer', default=u'Default Organizer'),
         description=_(
             u'help_default_organizer',
-            u"Default organizer to be used in events."),
+            u"Default organizer to be used in events.",
+        ),
         required=False,
         default='',
         vocabulary='plone.app.vocabularies.Catalog',
@@ -98,7 +89,7 @@ class IVenueSettings(Interface):
         pattern_options={
             'selectableTypes': ['Venue'],
             'basePath': get_base_path,
-        }
+        },
     )